--- python/runtime/utils/PyRemoteSimulatorQPU.cpp
@@ -1,26 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2022 - 2024 NVIDIA Corporation & Affiliates.                  *
- * All rights reserved.                                                        *
- *                                                                             *
- * This source code and the accompanying materials are made available under    *
- * the terms of the Apache License 2.0 which accompanies this distribution.    *
- ******************************************************************************/
-
-#include "common/BaseRemoteSimulatorQPU.h"
-
-using namespace mlir;
-
-namespace {
-
-// Remote QPU: delegating the execution to a remotely-hosted server, which can
-// reinstate the execution context and JIT-invoke the kernel.
-class PyRemoteSimulatorQPU : public cudaq::BaseRemoteSimulatorQPU {
-public:
-  PyRemoteSimulatorQPU() : BaseRemoteSimulatorQPU() {}
-
-  PyRemoteSimulatorQPU(PyRemoteSimulatorQPU &&) = delete;
-  virtual ~PyRemoteSimulatorQPU() = default;
-};
-} // namespace
-
-CUDAQ_REGISTER_TYPE(cudaq::QPU, PyRemoteSimulatorQPU, PyRemoteSimulatorQPU)
